# Improve performance of glearray.get() on Erlang